LayoutTests/fast/shadow-dom/caret-range-from-point-in-shadow-tree-expected.txt

bye
host: <i><br></i>, shadowRoot: <b><br></b>
range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 1, host.parentNode.offsetTop + 1)
PASS range.startContainer is host
PASS range.startOffset is 0
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

host: <i><br></i>, shadowRoot: <b><br></b>
range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 100, host.parentNode.offsetTop + 1)
PASS range.startContainer is host
PASS range.startOffset is 0
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

host: <i><br></i>, shadowRoot: hi
range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 1, host.parentNode.offsetTop + 1)
PASS range.startContainer is host
PASS range.startOffset is 0
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 100, host.parentNode.offsetTop + 1)
PASS range.startContainer is host
PASS range.startOffset is 0
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

host: <i><br></i>, shadowRoot: <slot></slot>
range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 1, host.parentNode.offsetTop + 1)
PASS range.startContainer is not null
PASS range.startContainer is host.querySelector("i")
PASS range.startOffset is 0
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 100, host.parentNode.offsetTop + 1)
PASS range.startContainer is not null
PASS range.startContainer is host.querySelector("i")
PASS range.startOffset is 0
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

host: hi, shadowRoot: <slot></slot>
range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 1, host.parentNode.offsetTop + 1)
PASS range.startContainer is host.firstChild
PASS range.startOffset is 0
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

host: hi, shadowRoot: <slot></slot>
range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 100, host.parentNode.offsetTop + 1)
PASS range.startContainer is host.firstChild
PASS range.startOffset is 2
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

host: bye, shadowRoot: <span id="inner-host"></span>, innerShadowRoot: hi
range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 1, host.parentNode.offsetTop + 1)
PASS range.startContainer is host
PASS range.startOffset is 0
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

host: bye, shadowRoot: <span id="inner-host"></span>, innerShadowRoot: hi
range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 100, host.parentNode.offsetTop + 1)
PASS range.startContainer is host
PASS range.startOffset is 0
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

host: bye, shadowRoot: <span id="inner-host" style="margin-left: 10px; border-left: solid 1px blue;">hi</span>, innerShadowRoot: <slot></slot>
range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 15, host.parentNode.offsetTop + 1)
PASS range.startContainer is host
PASS range.startOffset is 0
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

host: bye, shadowRoot: <span id="inner-host" style="margin-left: 10px; border-left: solid 1px blue;">hi</span>, innerShadowRoot: <slot></slot>
range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 100, host.parentNode.offsetTop + 1)
PASS range.startContainer is host
PASS range.startOffset is 0
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

host: bye, shadowRoot: <span id="inner-host" style="margin-left: 10px; border-left: solid 1px blue;"><slot></slot></span>, innerShadowRoot: <slot></slot>
range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 15, host.parentNode.offsetTop + 1)
PASS range.startContainer is host.firstChild
PASS range.startOffset is 0
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset

host: bye, shadowRoot: <span id="inner-host" style="margin-left: 10px; border-left: solid 1px blue;"><slot></slot></span>, innerShadowRoot: <slot></slot>
range = document.caretRangeFromPoint(host.parentNode.offsetLeft + 100, host.parentNode.offsetTop + 1)
PASS range.startContainer is host.firstChild
PASS range.startOffset is 3
PASS range.startContainer is range.endContainer
PASS range.startOffset is range.endOffset
PASS successfullyParsed is true

TEST COMPLETE
